Removed obsolete Providers
