+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submission-grid-row/SubmissionGridRow.tsx
@@ -7,8 +7,6 @@ import isNil from 'lodash/isNil';
 
 import { IPublicSubmission } from '../../../common/types';
 import { getContestsDetailsPageUrl } from '../../../common/urls/compose-client-urls';
-import { useUserProfileSubmissions } from '../../../hooks/submissions/use-profile-submissions';
-import { useProblems } from '../../../hooks/use-problems';
 import useTheme from '../../../hooks/use-theme';
 import { IAuthorizationReduxState } from '../../../redux/features/authorizationSlice';
 import { setProfile } from '../../../redux/features/usersSlice';
@@ -63,10 +61,8 @@ const SubmissionGridRow = ({
         testRuns,
     } = submission;
 
-    const { actions: { initiateRedirectionToProblem } } = useProblems();
     const { internalUser } =
         useSelector((reduxState: {authorization: IAuthorizationReduxState}) => reduxState.authorization);
-    const { actions: { getDecodedUsernameFromProfile } } = useUserProfileSubmissions();
     const dispatch = useAppDispatch();
 
     const [ competeIconAnchorElement, setCompeteIconAnchorElement ] = useState<HTMLElement | null>(null);
@@ -75,16 +71,16 @@ const SubmissionGridRow = ({
     const backgroundColorClassName = getColorClassName(themeColors.baseColor100);
 
     const usernameFromSubmission = isNil(user)
-        ? getDecodedUsernameFromProfile()
+        ? 'test'
         : user;
 
     const handleDetailsButtonSubmit = useCallback(
         () => {
             const submissionDetailsUrl = getSubmissionDetailsRedirectionUrl({ submissionId });
 
-            initiateRedirectionToProblem(problemId, submissionDetailsUrl);
+            navigate(submissionDetailsUrl);
         },
-        [ initiateRedirectionToProblem, problemId, submissionId ],
+        [ submissionId, navigate ],
     );
 
     const handleContestDetailsButtonSubmit = useCallback(
